requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
dkimpy==1.1.8
dnspython==2.7.0
checkdmarc==5.13.3
requests==2.32.3
tldextract==5.1.3
python-Levenshtein==0.26.1
marshmallow==3.23.2
pytest==8.3.4
gunicorn==22.0.0
